Use node executables in node_modules before $PATH (#363)
Many formatters are installed (or can be installed) via `npm`.  When
they are installed for a specific Node project, the executable is placed
in `$project_dir/node_modules/.bin` and is typically *not* on the
`$PATH`.  This change allows Neoformat to look for executables in the
`node_modules/.bin` directory before attempting to use the executable in
the `$PATH`.

For formatters that have `try_node_exe = 1`, and when
`g:neoformat_try_node_exe` is set to `1`, an upward search is performed
beginning at the current directory for `node_modules/.bin/$executable`.
If a match is found, it is used as the executable. If no match is found,
Neoformat will attempt to run the executable as-is.

As part of this change, I set `try_node_exe = 1` on all formatters that
made sense to me -- those related to JS or web development that appeared
to be available via NPM.

Note that this feature is off by default as `g:neoformat_try_node_exe`
must be set before it is used.

Fixes #143.