fix: add content type to deno fmt calls (#446)
Deno changed the default behavior for fmt calls using stdin content. By default it assumes all input is TypeScript and otherwise it needs to be specified by the --ext option. This adds the appropriate --ext flags for all file types associated with Deno (except markdown since that was added in an earlier).

Signed-off-by: StaticRocket <35777938+StaticRocket@users.noreply.github.com>